libobs: Rewrite macOS Hotkeys to use CGEventTap
Using CGEventTapCreate instead of addGlobalMonitorForEventsMatchingMask
to listen to global hotkeys has the advantage of tighter control over
what permissions we need. Since we only listen to and do not modify the
event, it is enough to have "Input Monitoring" instead of the full
"Accessibility" (which would allow modifying the event after catching
it).
It also might fix the issues we have with hotkeys randomly breaking down
after a while. I suspect that addGlobalMonitor could just be a wrapper
around CGEventTap and not handle timeouts.